test(config): refactor tests
